Make table cross-platform compatible
Due to the differences in allocation and other stuff between Windows and
Linux, the symbol list is not the same between the two, 'Dangan3Win.exe'
is missing on Linux. One solution is to scan for a unique AOB near the
'Dangan3Win.exe' symbol and register the symbol if not present.

This way, the base address for monocoins can be retrieved via the
pointer and offset approach which is more reliable in this case and
would work on both of the platforms.

Monocoins pointer address and offset are to be taken from https://fearlessrevolution.com/viewtopic.php?p=222782#p222782